Move TestBench API to the correct package
Fixes #9241
