--- uitest/src/test/java/com/vaadin/testbench/customelements/AbstractDateFieldElement.java
@@ -1,7 +1,0 @@
-package com.vaadin.testbench.customelements;
-
-import com.vaadin.testbench.elementsbase.ServerClass;
-
-@ServerClass("com.vaadin.ui.AbstractDateField")
-public class AbstractDateFieldElement extends DateFieldElement {
-}
