--- uitest/src/test/java/com/vaadin/testbench/customelements/LoginFormElement.java
@@ -1,9 +1,0 @@
-package com.vaadin.testbench.customelements;
-
-import com.vaadin.testbench.elements.AbstractSingleComponentContainerElement;
-import com.vaadin.testbench.elementsbase.ServerClass;
-
-@ServerClass("com.vaadin.ui.LoginForm")
-public class LoginFormElement extends AbstractSingleComponentContainerElement {
-
-}
