+++ uitest/src/test/java/com/vaadin/tests/components/datefield/DateFieldElementTest.java
@@ -5,7 +5,7 @@ import static org.hamcrest.core.Is.is;
 
 import org.junit.Test;
 
-import com.vaadin.testbench.customelements.AbstractDateFieldElement;
+import com.vaadin.testbench.elements.AbstractDateFieldElement;
 import com.vaadin.testbench.elements.DateFieldElement;
 import com.vaadin.tests.tb3.SingleBrowserTest;
 
@@ -15,7 +15,7 @@ public class DateFieldElementTest extends SingleBrowserTest {
     public void dateFieldElementIsLocated() {
         openTestURL();
 
-        assertThat($(DateFieldElement.class).all().size(), is(1));
+        assertThat($(DateFieldElement.class).all().size(), is(2));
         assertThat($(AbstractDateFieldElement.class).all().size(), is(2));
     }
