+++ uitest/src/test/java/com/vaadin/tests/components/uitest/ThemeTest.java
@@ -8,9 +8,9 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 
 import com.vaadin.testbench.By;
-import com.vaadin.testbench.customelements.AbstractDateFieldElement;
 import com.vaadin.testbench.elements.ButtonElement;
 import com.vaadin.testbench.elements.ComboBoxElement;
+import com.vaadin.testbench.elements.DateFieldElement;
 import com.vaadin.testbench.elements.NotificationElement;
 import com.vaadin.testbench.elements.TabSheetElement;
 import com.vaadin.testbench.elements.TableElement;
@@ -187,13 +187,13 @@ public abstract class ThemeTest extends MultiBrowserTest {
 
     private void testDates() throws IOException {
         compareScreen("dates");
-        $(AbstractDateFieldElement.class).id("datefield0").openPopup();
+        $(DateFieldElement.class).id("datefield0").openPopup();
         compareScreen("dates-first-popup");
-        $(AbstractDateFieldElement.class).id("datefield1").openPopup();
+        $(DateFieldElement.class).id("datefield1").openPopup();
         compareScreen("dates-second-popup");
-        $(AbstractDateFieldElement.class).id("datefield2").openPopup();
+        $(DateFieldElement.class).id("datefield2").openPopup();
         compareScreen("dates-third-popup");
-        $(AbstractDateFieldElement.class).id("datefield3").openPopup();
+        $(DateFieldElement.class).id("datefield3").openPopup();
         compareScreen("dates-fourth-popup");
     }
 }
